Don't remove staging

+++ buildroot-scripts/prepare-cache.sh
@@ -2,6 +2,7 @@
 set -e
 
 BR="$1"
+REMOVE_STAGING="$2"
 
 if ! [[ -d "$BR" ]]; then
   echo "Buildroot folder does not exist."
@@ -30,11 +31,17 @@ if [[ -d "$OUTPUT" ]]; then
   rm -rf target images
   mkdir -p target images
 
-  # Remove staging dir
-  STAGING_DIR=$(readlink -f staging)
-  rm -rf "$STAGING_DIR"
-  mkdir -p "$STAGING_DIR"
+  if [[ -n "$REMOVE_STAGING" ]]; then
+    # Remove staging dir
+    STAGING_DIR=$(readlink -f staging)
+    rm -rf "$STAGING_DIR"
+    mkdir -p "$STAGING_DIR"
+  fi
 fi
 
 # Remove all stamps for installing images, initramfs, target and staging
-rm -f $(find . -name ".stamp_images_installed" -or -name ".stamp_initramfs_rebuilt" -or -name ".stamp_target_installed" -or -name ".stamp_staging_installed")
+rm -f $(find . -name ".stamp_images_installed" -or -name ".stamp_initramfs_rebuilt" -or -name ".stamp_target_installed")
+
+if [[ -n "$REMOVE_STAGING" ]]; then
+  rm -f $(find . -name ".stamp_staging_installed")
+fi
